commit 13 update library, migrate kotlin 6, fix bug. versionCode 2 versionName "1.1"

+++ app/src/main/java/com/appsforlife/dictaphone/viewModels/RemoveViewModel.kt
@@ -18,7 +18,7 @@ class RemoveViewModel(
     private var job = Job()
     private val uiScope = CoroutineScope(Dispatchers.Main + job)
 
-    fun removeRecord(recordId: Long) {
+    fun removeRecord(recordId: Int) {
         recordDAO = RecordDB.getInstance(application).recordDAO
 
         try {
